Move configure scripts to CLI packages to fix compilation issues
