+++ workspaces/templates-lib/packages/template-backup-central-cli/package.json
@@ -40,9 +40,11 @@
     "patch-when-changed": "git diff --quiet HEAD^...HEAD -- . && git diff --quiet -- . && git diff --quiet --staged -- .  || yarn version patch"
   },
   "dependencies": {
+    "@goldstack/infra": "0.4.41",
     "@goldstack/template-backup-central": "0.1.3",
     "@goldstack/utils-cli": "0.3.32",
     "@goldstack/utils-package": "0.4.43",
+    "@goldstack/utils-sh": "0.5.39",
     "@goldstack/utils-terraform": "0.4.80",
     "@goldstack/utils-terraform-aws": "0.4.85",
     "source-map-support": "^0.5.21",
