+++ workspaces/templates/packages/backup-central/tsconfig.json
@@ -2,6 +2,7 @@
   "extends": "../../tsconfig.base.json",
   "include": [
     "./src/**/*",
+    "./scripts/**/*",
     "./schemas/**/*.json",
     "./src/**/*.json",
     "./goldstack.json"
@@ -9,7 +10,7 @@
   "compilerOptions": {
     "outDir": "dist",
     "baseUrl": ".",
-    "types": []
+    "types": ["node"]
   },
   "references": [
     {
